// api-client/models/create-asset-dto.ts
/* tslint:disable */
 
/**
 * API Documentation
 * API - IT Asset Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateAssetDto
 */
export interface CreateAssetDto {
    /**
     * Name of the asset
     * @type {string}
     * @memberof CreateAssetDto
     */
    'name': string;
    /**
     * Unique stock keeping unit (SKU) code for the asset
     * @type {string}
     * @memberof CreateAssetDto
     */
    'sku': string;
    /**
     * Description of the asset
     * @type {string}
     * @memberof CreateAssetDto
     */
    'description': string;
    /**
     * Model of the asset
     * @type {string}
     * @memberof CreateAssetDto
     */
    'model': string;
    /**
     * Brand of the asset
     * @type {string}
     * @memberof CreateAssetDto
     */
    'brand': string;
    /**
     * ID of the category this asset belongs to
     * @type {number}
     * @memberof CreateAssetDto
     */
    'categoryId': number;
    /**
     * ID of the employee assigned as custodian of this asset
     * @type {number}
     * @memberof CreateAssetDto
     */
    'employeeId': number;
}