// api-client/models/update-asset-dto.ts
/* tslint:disable */
 
/**
 * API Documentation
 * API - IT Asset Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateAssetDto
 */
export interface UpdateAssetDto {
    /**
     * Name of the asset
     * @type {string}
     * @memberof UpdateAssetDto
     */
    'name'?: string;
    /**
     * Unique stock keeping unit (SKU) code for the asset
     * @type {string}
     * @memberof UpdateAssetDto
     */
    'sku'?: string;
    /**
     * Description of the asset
     * @type {string}
     * @memberof UpdateAssetDto
     */
    'description'?: string;
    /**
     * Model of the asset
     * @type {string}
     * @memberof UpdateAssetDto
     */
    'model'?: string;
    /**
     * Brand of the asset
     * @type {string}
     * @memberof UpdateAssetDto
     */
    'brand'?: string;
    /**
     * ID of the category this asset belongs to
     * @type {number}
     * @memberof UpdateAssetDto
     */
    'categoryId'?: number;
    /**
     * ID of the employee assigned as custodian of this asset
     * @type {number}
     * @memberof UpdateAssetDto
     */
    'employeeId'?: number;
}